+++ packages/fiori/src/bundle.esm.ts
@@ -12,7 +12,6 @@ import "./features/CoPilotAnimation.js";
 import "./illustrations/AllIllustrations.js";
 
 // FIORI components
-import Bar from "./Bar.js";
 import BarcodeScannerDialog from "./BarcodeScannerDialog.js";
 import DynamicSideContent from "./DynamicSideContent.js";
 import FilterItem from "./FilterItem.js";
