+++ packages/fiori/src/themes/sap_horizon_hcw_exp/parameters-bundle.css
@@ -1,5 +1,3 @@
-@import "../base/sizes-parameters.css";
-@import "../base/Bar-parameters.css";
 @import "./FlexibleColumnLayout-parameters.css";
 @import "../base/IllustratedMessage-parameters.css";
 @import "./MediaGallery-parameters.css";
