+++ packages/main/src/Bar.hbs
