+++ packages/main/src/themes/base/sizes-parameters.css
@@ -153,6 +153,9 @@
 [data-ui5-compact-size],
 .ui5-content-density-compact,
 .sapUiSizeCompact {
+	--_ui5_bar_base_height: 2.5rem;
+	--_ui5_bar_subheader_height: 2.25rem;
+
 	--_ui5_button_base_height: var(--sapElement_Compact_Height);
 	--_ui5_button_base_padding: 0.4375rem;
 	--_ui5_button_base_min_width: 2rem;
