+++ packages/main/src/themes/sap_belize/parameters-bundle.css
@@ -1,6 +1,7 @@
 @import "../base/Avatar-parameters.css";
 @import "../base/AvatarGroup-parameters.css";
 @import "../base/Badge-parameters.css";
+@import "../base/Bar-parameters.css";
 @import "./BrowserScrollbar-parameters.css";
 @import "./BusyIndicator-parameters.css";
 @import "../base/CalendarLegend-parameters.css";
