+++ packages/main/src/themes/sap_fiori_3_dark/parameters-bundle.css
@@ -1,6 +1,7 @@
 @import "./Avatar-parameters.css";
 @import "../base/AvatarGroup-parameters.css";
 @import "./Badge-parameters.css";
+@import "../base/Bar-parameters.css";
 @import "../base/BrowserScrollbar-parameters.css";
 @import "../base/BusyIndicator-parameters.css";
 @import "./Button-parameters.css";
