+++ packages/main/src/themes/sap_fiori_3_hcb/Bar-parameters.css
