+++ packages/main/src/themes/sap_horizon/parameters-bundle.css
@@ -1,6 +1,7 @@
 @import "./Avatar-parameters.css";
 @import "./AvatarGroup-parameters.css";
 @import "./Badge-parameters.css";
+@import "../base/Bar-parameters.css";
 @import "./Breadcrumbs-parameters.css";
 @import "../base/BrowserScrollbar-parameters.css";
 @import "./BusyIndicator-parameters.css";
