+++ packages/main/src/themes/sap_horizon_hcb/Bar-parameters.css
