+++ packages/main/src/themes/sap_horizon_hcw/Bar-parameters.css
