+++ packages/main/src/types/BarDesign.ts
