+++ packages/website/docs/_components_pages/main/Bar.mdx
@@ -1,4 +1,4 @@
-import Basic from "../../_samples/fiori/Bar/Basic/Basic.md";
+import Basic from "../../_samples/main/Bar/Basic/Basic.md";
 
 <%COMPONENT_OVERVIEW%>
