+++ packages/website/docs/_samples/main/Bar/Basic/main.js
@@ -1,5 +1,5 @@
 import "@ui5/webcomponents/dist/Button.js";
 import "@ui5/webcomponents/dist/Label.js";
-import "@ui5/webcomponents-fiori/dist/Bar.js";
+import "@ui5/webcomponents/dist/Bar.js";
 import "@ui5/webcomponents-icons/dist/home.js"
 import "@ui5/webcomponents-icons/dist/action-settings.js"
